pyproject.toml
[project]
name = "batch-image-optimiser"
version = "0.1.0"
description = "A command-line tool for batch image optimisation and conversion."
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.0",
    "pillow>=12.0.0",
]
